moves validators from tools into lib
